Add files via upload
So when i learned a bit of css, i made this sliding text display for my friend. Now, i know you can make normal sliding text using css too. But i made this to understand that you can make a ton of animations (331 to be accurate) in sync to make it look like a single animation, it can be used to explore different possibilities to use in someone else's projects. However, this was my first css project. All of the code is original and is written by me.